+++ tools/pbespgsolve/ParityGame_pbes.h
@@ -19,8 +19,8 @@ void ParityGame::assign_pbes( mcrl2::pbes_system::pbes<Container> &pbes,
     verti begin = 0, end = 3;
     for (verti v = begin; v < end; ++v)
     {
-        std::set<unsigned> deps = pgg.get_dependencies(v);
-        for ( std::set<unsigned>::const_iterator it = deps.begin();
+        std::set<size_t> deps = pgg.get_dependencies(v);
+        for ( std::set<size_t>::const_iterator it = deps.begin();
                 it != deps.end(); ++it )
         {
             verti w = (verti)*it;
